+++ tests/test_settings_migrations.py
@@ -1,9 +1,12 @@
 import unittest
 
-from settings_migrations import add_service_to_serialized_settings
+from settings_migrations import PINTEREST_DEFAULT_MIGRATION, add_service_to_serialized_settings
 
 
 class SettingsMigrationTests(unittest.TestCase):
+    def test_pinterest_migration_has_a_stable_unique_name(self):
+        self.assertEqual(PINTEREST_DEFAULT_MIGRATION, "enable_pinterest_pins_v1")
+
     def test_new_service_is_appended_to_existing_settings(self):
         serialized = repr(["Twitter", "Instagram"])